Use postcss to extract styles in Expo example
Extract styles to static CSS using postcss, as there is no existing
Metro integration.

Fix #34
Ref #236
